Move C interface functions to separate source file
Refactored the C interface functions from MainPluginContext.cpp into a new file, MainPluginContext_c.cpp, and updated CMakeLists.txt to include the new source file. This improves code organization by separating C linkage functions from the main C++ implementation.
