+++ CMakeLists.txt
@@ -47,7 +47,7 @@ target_compile_definitions(
   PRIVATE PLUGIN_NAME="${CMAKE_PROJECT_NAME}" PLUGIN_VERSION="${CMAKE_PROJECT_VERSION}"
 )
 
-target_sources(${CMAKE_PROJECT_NAME} PRIVATE src/MainPluginContext.cpp src/plugin-main.c)
+target_sources(${CMAKE_PROJECT_NAME} PRIVATE src/MainPluginContext_c.cpp src/MainPluginContext.cpp src/plugin-main.c)
 target_link_libraries(
   ${CMAKE_PROJECT_NAME}
   PUBLIC OBS::libobs cpr::cpr Eigen3::Eigen fmt::fmt ncnn opencv_core opencv_imgproc opencv_imgcodecs semver::semver
